refactor and added styling
