+++ client/src/Components/AddEquipmentButtons/index.js
@@ -1,0 +1,3 @@
+import AddEquipmentButtons from "./AddEquipmentButtons";
+
+export default AddEquipmentButtons;
