+++ client/src/Components/AddSnowboardModal/index.js
@@ -1,0 +1,3 @@
+import AddSnowboardModal from "./AddSnowboardModal";
+
+export default AddSnowboardModal;
