+++ client/src/Components/UpdateEquipmentButtons/index.js
@@ -1,0 +1,3 @@
+import UpdateEquipmentButtons from "./UpdateEquipmentButtons";
+
+export default UpdateEquipmentButtons;
