+++ client/src/Components/index.js
@@ -1,15 +1,31 @@
 import UserSearchForm from "./UserSearchForm";
 import EquipmentSearch from "./EquipmentSearch";
-import EquipmentList from './EquipmentList';
+import EquipmentList from "./EquipmentList";
 import EquipmentSearchBar from "./EquipmentSearchBar";
 import ContractSubmit from "./ContractSubmit";
 import ContractDateSelection from "./ContractDateSelection";
+import AddBootModal from "./AddBootModal";
+import AddSkiModal from "./AddSkiModal";
+import AddSnowboardModal from "./AddSnowboardModal";
+import AddEquipmentButtons from "./AddEquipmentButtons";
+import UpdateBootModal from "./UpdateBootModal";
+import UpdateSkiModal from "./UpdateSkiModal";
+import UpdateSnowboardModal from "./UpdateSnowboardModal";
+import UpdateEquipmentButtons from "./UpdateEquipmentButtons";
 
-export { 
-    UserSearchForm, 
-    EquipmentSearch, 
-    EquipmentSearchBar, 
-    EquipmentList, 
-    ContractSubmit,
-    ContractDateSelection
-}
+export {
+  UserSearchForm,
+  EquipmentSearch,
+  EquipmentSearchBar,
+  EquipmentList,
+  ContractSubmit,
+  ContractDateSelection,
+  AddBootModal,
+  AddSkiModal,
+  AddSnowboardModal,
+  AddEquipmentButtons,
+  UpdateBootModal,
+  UpdateSkiModal,
+  UpdateSnowboardModal,
+  UpdateEquipmentButtons,
+};
