+++ client/src/equipment.css
@@ -1,5 +1,6 @@
 #equipmentCard {
-    margin: auto;
-    width: 60%;
-    align-items: center;
+  margin: auto;
+  width: 50%;
+  align-items: center;
+  border: none;
 }
